The compiler is PsyQ 4.5, not 4.6 — 227 functions
With the same 220 sources, 4.5's cc1psx builds a byte-identical binary and
4.6 does not, and no function matches under 4.6 that fails under 4.5.
Switching then unlocked eight parked functions at once: func_800495A4,
func_80024954, func_80038388, func_800383B0, func_8007368C, func_800598E4,
func_80042874, func_8003FCD8. Those were three separately-named park
classes -- the coupled-ordering wrappers, the $a0-as-scratch group, and
assorted register-allocation misses. All one thing: the wrong compiler.

Found via func_800495A4, whose flags are pinned by two matching neighbours
0x10 and 0x38 away. Under 4.6 those flags give retail's registers with the
wrong ordering, and -O2 gives the ordering with the wrong registers. The
measurement that closed it: -O1 -G0 -fschedule-insns2 emits output
byte-identical to -O2 -G0, so enabling the post-reload scheduler changes
register numbers too -- ordering and allocation are one switch and 4.6
cannot produce retail's combination. 4.5 produces it directly.

Cost: 4.5 narrows a u8 return with an extra andi where 4.6 does not, which
broke func_8007058C and func_8003B7E0. Both fixed by returning s32, since
the value comes from an lbu and is already zero-extended. Point-of-use
casting applied to return types.

"Confirmed twice" was not confirmation. Ghidra's detection is a heuristic
over library signatures the two versions share, and decomp.me's config is
someone else's guess. And the byte-identical build was not evidence either:
219 functions matched under the wrong compiler because most are too small
for the versions to differ. A passing global check validates an assumption
only where it was exercised -- the way to test one is to find the single
discriminating case, not to count successes elsewhere.

Still failing under 4.5, now honest unknowns: func_80035DF4, func_800428EC,
func_80049200, func_8005B5FC (wrong size), func_8003F7D4, func_80070988,
func_80071424, func_80071460 (differing). Anything recorded about the parked
set before this commit was measured under the wrong compiler.

build.py defaults to 4.5; YGOFM_PSYQ=46 selects 4.6 for comparison, and the
flag stamp records which compiler built each object. Also fixed the differ
list truncating at 25, which cost two wrong readings during this
investigation -- YGOFM_ALL_DIFFS=1 prints all.

+++ src/func_8003B7E0.c
@@ -1,9 +1,9 @@
 #include "common.h"
 
-u8 func_8003B7E0(u8 *arg0) {
+s32 func_8003B7E0(u8 *arg0) {
     u8 **pp = (u8 **)(arg0 + *(s8 *)(arg0 + 0x58) * 4);
     u8 *p = *pp;
-    u8 v = *p;
+    s32 v = *p;
 
     *pp = p + 1;
     return v;
