+++ src/func_8007058C.c
@@ -1,8 +1,8 @@
 #include "common.h"
 
-u8 func_8007058C(void) {
+s32 func_8007058C(void) {
     u8 *p = D_800F5BE8.cursor;
-    u8 v = *p;
+    s32 v = *p;
 
     D_800F5BE8.cursor = p + 1;
     return v;
